extract hibernate specific deadlock detection to DeadLockDetector interface
